requirements.txt
azure-functions
ga4gh.vrs[extras]==0.8.5